+++ packages/cozy-stack-client/src/DocumentCollection.spec.js
@@ -1,7 +1,7 @@
-jest.mock('../CozyStackClient')
+jest.mock('./CozyStackClient')
 
-import CozyStackClient from '../CozyStackClient'
-import DocumentCollection from '../DocumentCollection'
+import CozyStackClient from './CozyStackClient'
+import DocumentCollection from './DocumentCollection'
 
 const ALL_RESPONSE_FIXTURE = {
   offset: 0,
@@ -251,14 +251,14 @@ describe('DocumentCollection', () => {
 
     it('should accept a sort option', async () => {
       const collection = new DocumentCollection('io.cozy.todos', client)
-      await collection.find({ done: false }, { sort: { label: 'desc' } })
+      await collection.find({ done: false }, { sort: [{ label: 'desc' }] })
       expect(client.fetchJSON).toHaveBeenLastCalledWith(
         'POST',
         '/data/io.cozy.todos/_find',
         {
           skip: 0,
           selector: { done: false },
-          sort: [{ done: 'desc' }, { label: 'desc' }],
+          sort: [{ label: 'desc' }, { done: 'desc' }],
           use_index: '_design/123456'
         }
       )
@@ -266,25 +266,25 @@ describe('DocumentCollection', () => {
 
     it('should use a valid default sorting option', async () => {
       const collection = new DocumentCollection('io.cozy.todos', client)
-      await collection.find({ done: false }, { sort: { label: 'asc' } })
+      await collection.find({ done: false }, { sort: [{ label: 'asc' }] })
       expect(client.fetchJSON).toHaveBeenLastCalledWith(
         'POST',
         '/data/io.cozy.todos/_find',
         {
           skip: 0,
           selector: { done: false },
-          sort: [{ done: 'asc' }, { label: 'asc' }],
+          sort: [{ label: 'asc' }, { done: 'asc' }],
           use_index: '_design/123456'
         }
       )
-      await collection.find({ done: false }, { sort: { label: 'desc' } })
+      await collection.find({ done: false }, { sort: [{ label: 'desc' }] })
       expect(client.fetchJSON).toHaveBeenLastCalledWith(
         'POST',
         '/data/io.cozy.todos/_find',
         {
           skip: 0,
           selector: { done: false },
-          sort: [{ done: 'desc' }, { label: 'desc' }],
+          sort: [{ label: 'desc' }, { done: 'desc' }],
           use_index: '_design/123456'
         }
       )
@@ -295,9 +295,27 @@ describe('DocumentCollection', () => {
       await expect(
         collection.find(
           { done: false },
-          { sort: { label: 'asc', _id: 'desc' } }
+          { sort: [{ label: 'asc' }, { _id: 'desc' }] }
         )
       ).rejects.toThrow()
+    })
+
+    it('should respect the sorting order', async () => {
+      const collection = new DocumentCollection('io.cozy.todos', client)
+      await collection.find(
+        { done: false },
+        { sort: [{ label: 'desc' }, { _id: 'desc' }] }
+      )
+      expect(client.fetchJSON).toHaveBeenLastCalledWith(
+        'POST',
+        '/data/io.cozy.todos/_find',
+        {
+          skip: 0,
+          selector: { done: false },
+          sort: [{ label: 'desc' }, { _id: 'desc' }, { done: 'desc' }],
+          use_index: '_design/123456'
+        }
+      )
     })
 
     it('should return a correct JSON API response', async () => {